Move WorldEventEvents.RENDER to separate class
Fixes crash on dedicated server
